backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
daphne==4.0.0
mysqlclient==2.2.1
python-dotenv==1.0.0
cryptography==42.0.0
paramiko==3.4.0
celery==5.3.6
redis==5.0.1
channels==4.0.0
channels-redis==4.2.0
django-celery-beat==2.6.0
pyotp==2.9.0
qrcode==7.4.2
pillow==10.2.0
django-auth-ldap==4.6.0
python-ldap==3.4.4
requests==2.31.0
openpyxl==3.1.2
reportlab==4.0.9
